requirements.txt
# Project Revelare - Python Dependencies
# Core web framework
Flask>=2.3.0,<3.0.0

# HTTP requests for external APIs
requests>=2.28.0,<3.0.0
py7zr>=0.20.0

# Image processing and fractal generation
Pillow>=9.0.0,<11.0.0
numpy>=1.21.0,<2.0.0
exifread>=3.0.0

# Document processing
pypdf>=6.0.0
python-docx>=0.8.11
pandas>=1.5.0

# GeoIP database support
maxminddb>=3.0.0

# Optional OCR functionality (uncomment if needed)
# pytesseract>=0.3.10
# opencv-python>=4.5.0

# Optional transcription functionality (uncomment if needed)
# openai-whisper>=20231117
# SpeechRecognition>=3.10.0
# pydub>=0.25.1

# Optional data analysis (uncomment if needed)
# pandas>=1.5.0
# matplotlib>=3.5.0
# seaborn>=0.11.0

# Optional database support (uncomment if needed)
# sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymongo>=4.0.0  # MongoDB

# Optional web scraping (uncomment if needed)
# beautifulsoup4>=4.11.0
# scrapy>=2.6.0
# selenium>=4.0.0

# Optional security and cryptography (uncomment if needed)
# cryptography>=3.4.8
# pycryptodome>=3.15.0

# Optional configuration management (uncomment if needed)
# python-dotenv>=0.19.0

# Optional OAuth support (uncomment if needed)
# requests-oauthlib>=1.3.0

# Development and testing (uncomment if needed)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991

# Production deployment (uncomment if needed)
# gunicorn>=20.1.0
# uvicorn>=0.18.0
# celery>=5.2.0
# redis>=4.3.0

# Note: This project primarily uses Python standard library modules
# The above packages are only required for specific optional features
# Core functionality works with just Flask and requests
